Fix. fix deleteDatabase controller function

+++ src/controllers/database.controller.js
@@ -2,6 +2,7 @@
 const User = require("../models/User");
 const { Database } = require("../models/Database");
 const { Field } = require("../models/Field");
+const { Document } = require("../models/Document");
 
 exports.getAllDatabases = async function (req, res, next) {
   const userId = req.params.userid;
@@ -21,39 +22,6 @@ exports.getAllDatabases = async function (req, res, next) {
   } catch (err) {
     console.error("Error while fetching databases", err);
     res.status(500).json({ error: "Failed to retrieve databases" });
-  }
-};
-
-exports.createDatabase = async function (req, res, next) {
-  const userId = req.params.userid;
-
-  try {
-    const user = await User.findById(userId);
-
-    if (!user) {
-      return res.status(404).json({ error: "User Not Found" });
-    }
-
-    const newDatabase = await Database.create({
-      name: req.body.dbName,
-      createdBy: userId,
-    });
-
-    await Promise.all(
-      req.body.fields.map(async (item) => {
-        const field = await Field.create({ name: item.name, type: item.type });
-        newDatabase.fields.push(field);
-
-        return field;
-      }),
-    );
-
-    await user.save();
-
-    res.status(201).json({ newDatabase, user });
-  } catch (error) {
-    console.error("Error while fetching database", error);
-    res.status(500).json({ error: "Failed to retrieve database" });
   }
 };
 
@@ -113,10 +81,35 @@ exports.getDatabase = async function (req, res, next) {
 };
 
 exports.deleteDatabase = async function (req, res, next) {
+  const userId = req.params.userid;
   const databaseId = req.params.databaseid;
 
   try {
-    await Database.findByIdAndDelete(databaseId);
+    const user = await User.findById(userId);
+
+    if (!user) {
+      return res.status(404).json({ error: "User Not Found" });
+    }
+
+    const deletedDatabase = await Database.findByIdAndDelete(databaseId);
+
+    const documentIds = [];
+    const fieldIds = [];
+
+    deletedDatabase.documents.forEach(function (document) {
+      documentIds.push(document._id);
+    });
+
+    deletedDatabase.fields.forEach(function (field) {
+      fieldIds.push(field._id);
+    });
+
+    await Document.deleteMany({ _id: { $in: documentIds } });
+    await Field.deleteMany({ _id: { $in: fieldIds } });
+
+    user.databases = user.databases.filter(function (database) {
+      return database.id !== databaseId;
+    });
 
     res.status(200).json("Database successfully deleted");
   } catch (err) {
